Code cleanup and Readme

+++ VantageLibrary/Types/ResponseTypes/VantageDomainLoad.cs
@@ -2,5 +2,5 @@
 
 public class VantageDomainLoad
 {
-    public long Load { get; set; }
+    public int Load { get; set; }
 }
