+++ VantageLibrary/Types/ResponseTypes/VantageWorkflowJobInputs.cs
@@ -11,6 +11,6 @@ public class VantageWorkflowJobInputs {
     public string JobName { get; set; }
     public List<VantageLabel> Labels { get; set; }
     public List<VantageMedia> Medias { get; set; }
-    public long Priority { get; set; }
+    public int Priority { get; set; }
     public List<VantageVariable> Variables { get; set; }
 }
